% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkr.R
\name{chunkr}
\alias{chunkr}
\title{chunkr splits a dataset into chunks}
\usage{
chunkr(mydf, nchunks = 1L, ...)
}
\arguments{
\item{mydf}{the dataframe to be used.}

\item{nchunks}{an integer target number of chunks in which the dataframe is to be split (1 <= nchunks <= N) where N is the number of units.}

\item{...}{a list of one or more variables that defines the units which should not be split apart.}
}
\value{

}
\description{
splits a dataset into a specified number of chunks so that they can be processed sequentially
}
\details{
The splitting is done in terms of units defined by the combinations of the levels of one or more variables.
Any output files will be produced in .rds format.
}
\examples{
mydf <- mtcars
nchunks <- chunkr(mydf, nchunks = 100L, "gear")
nchunks


}
