% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitizr.R
\name{digitizr}
\alias{digitizr}
\title{a wrapper for the magick package, focussed on extracting data from images containing graphs}
\usage{
digitizr(
  inputfilename = NULL,
  outputfilename = NULL,
  xvarname = NULL,
  yvarname = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  yscalemax = NULL,
  threshold = NULL,
  mygraphsize = 1000
)
}
\arguments{
\item{inputfilename}{Name of the input image file. This file should contain only the graph area, not the axes, titles or legends.}

\item{outputfilename}{Name to be used for the output image file (PNG format).}

\item{xvarname}{Name to be used for the x variable.}

\item{yvarname}{Name to be used for the y variable.}

\item{xmin}{Minimum x value.}

\item{xmax}{Maximum x value.}

\item{ymin}{Minimum y value.}

\item{ymax}{Maximum y value.}

\item{yscalemax}{Maximum value to be used for the y-scale.}

\item{threshold}{The threshold is a percentage written as a character string, e.g. "70\%".  Some experimentation might be needed to find the value that works best.}

\item{mygraphsize}{The graph size in pixels.}
}
\value{

}
\description{
a wrapper for the magick package, focussed on extracting data from images containing graphs
}
\examples{
myinputfile <- system.file("extdata", "ari.png", package = "humblr")
extracted_data <- digitizr(inputfilename=myinputfile, outputfilename="extracted_data_graph.png", xvarname="año", yvarname="millones de pasajeros", xmin=2020, xmax=2050, ymin=0.928253, ymax=4.159696, yscalemax=5, threshold="70\%")

}
