% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepr.R
\name{keepr}
\alias{keepr}
\title{keeps a specified list of variables and drops the rest}
\usage{
keepr(mydf, ...)
}
\arguments{
\item{mydf}{the dataframe to be modified.}

\item{...}{a list of the variables to be kept.}
}
\value{

}
\description{
keepr is used to keep a list of variables in a dataframe (and to drop the rest).  Inspired by the "keep" command in Stata.
}
\details{
keepr is used internally by the fsm package.
}
\examples{
mydf <- mtcars
mydf <- keepr(mydf, "mpg", "cyl", "displ")
mydf

}
