% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs_sampling.R
\name{lhs_sampling}
\alias{lhs_sampling}
\title{generates a Latin hypercube based on inputs provided in an Excel spreadsheet.}
\usage{
lhs_sampling(
  myfilename = NULL,
  mytestspervariable = 10,
  myseed = 12345L,
  mymaxsweeps = 4,
  myeps = 0.01,
  mydup = 5,
  mypop = 1000,
  mygen = 8,
  mypmut = 0.1,
  mygraphsize = 1000,
  mypch = 19,
  mycol = "blue",
  mycex = 0.5
)
}
\arguments{
\item{myfilename}{is the name of the input Excel file}

\item{mytestspervariable}{the number of tests per variable in the input file}

\item{myseed}{the random number seed to be used}

\item{mymaxsweeps}{see lhs package documentation}

\item{myeps}{see lhs package documentation}

\item{mydup}{see lhs package documentation}

\item{mypop}{see lhs package documentation}

\item{mygen}{see lhs package documentation}

\item{mypmut}{see lhs package documentation}

\item{mygraphsize}{pixel dimension to be used for scatter graph matrices. see "width" and "height" in png {grDevices}.}

\item{mypch}{see "pch" in par {graphics}}

\item{mycol}{see "col" in par {graphics}}

\item{mycex}{see "cex" in par {graphics}}
}
\value{

}
\description{
facilitates the generation of a Latin hypercubes for a set of variables based on inputs specified in a spreadsheet table.
}
\details{
The spreadsheet should have only one sheet with the following columns and one row for each variable to be simulated:
\itemize{
\item Index (numeric)
\item Variable (character)
\item codename (character)
\item Base (numeric)
\item Minimum (numeric)
\item Mode (numeric)
\item Maximum (numeric)
\item Distribution (character) - Uniform, Triangular, or PERT.
\item Shape (numeric) - only needed if distribution is "PERT"
}
}
\examples{
myinputfile <- system.file("extdata", "Risk variable distributions for Latin Hypercube sample.xlsx", package = "humblr")
mytest1 <- lhs_sampling(myinputfile, mytestspervariable = 10, myseed = 12345L, mymaxsweeps = 4, myeps = 0.01, mydup = 5, mypop = 1000, mygen = 8, mypmut = 0.1, mygraphsize = 1000, mypch = 19, mycol = "blue", mycex = 0.5)
mytest1

}
