% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmodelr.R
\name{mcmodelr}
\alias{mcmodelr}
\title{monte carlo simulation function for use with mcr and model_ols_gpm}
\usage{
mcmodelr(
  nsims = NULL,
  myinputspreadsheet = NULL,
  myinputdatasheet = NULL,
  mypredictiondf = NULL,
  myresponse_ols = NULL,
  myresponse_gpm = NULL,
  myterms = NULL,
  sigmafrom = NULL,
  sigmato = NULL,
  sigmaby = NULL,
  myfolds = NULL,
  myseed1 = NULL,
  myseed2 = NULL,
  mygraphsize = 1000,
  mygraphtextsize = 10,
  mytag = NULL
)
}
\arguments{
\item{nsims}{the desired number of Monte Carlo simulations}

\item{myinputspreadsheet}{is the name of the spreadsheet with the input data.}

\item{myinputdatasheet}{the name of the sheet with the input data.}

\item{mypredictiondf}{(optional) a dataframe to be used as a basis for predictions}

\item{myresponse_ols}{the name of the dependent variable for the OLS model}

\item{myresponse_gpm}{the name of the dependent variable for the GPM model.  This will normally be the residuals of the OLS model: "residuals".}

\item{myterms}{a character vector containing the names of the independent variables.}

\item{sigmafrom}{the lower limit of the test range for sigma}

\item{sigmato}{the upper limit of the test range for sigma}

\item{sigmaby}{the increment of the test values for sigma}

\item{myfolds}{the number of folds for k-fold cross-validation}

\item{myseed1}{the first random number seed}

\item{myseed2}{the second random number seed}

\item{mygraphsize}{the pixel dimension for the output graphs}

\item{mygraphtextsize}{the size of all text elements on the graph.}

\item{mytag}{a character string that will be used in the output file names.  #'}
}
\value{

}
\description{
takes model_ols_gpm, passes it input data produced with mcr, and passes back key results
}
\examples{

nsims <- 1000

myinputfile <- system.file("extdata", "risk_variable_distributions.xlsx", package = "humblr")
mytest1 <- mcr(myinputfile, nsims = nsims, myseed = 12345L)
mytest1

mycorrmatfile <- system.file("extdata", "correlation_matrix.xlsx", package = "humblr")
mytest2 <- mcr(myinputfile, mycorrmatfilename = mycorrmatfile, nsims = nsims, myseed = 12345L)
mytest2

myinputfile <- system.file("extdata", "model_ols_gpm_test.xlsx", package = "humblr")
mymodel_results <- mcmodelr(nsims = nsims, myinputspreadsheet = myinputfile, myinputdatasheet = "data", mypredictiondf = mytest2, myresponse_ols = "depvar", myresponse_gpm = "residuals", myterms = c("var2",	"var3",	"var4",	"var5",	"var6",	"var7",	"var8",	"var9",	"var10"), sigmafrom = 1, sigmato= 10, sigmaby = 1, myfolds = 10, myseed1 = 123, myseed2 = 456, mygraphsize = 1000, mygraphtextsize = 20, mytag = "mymodel")



}
