% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gpm.R
\name{model_gpm}
\alias{model_gpm}
\title{Gaussian Process Model.}
\usage{
model_gpm(
  mydf = NULL,
  myresponse = NULL,
  myterms = NULL,
  sigmafrom = 0.01,
  sigmato = 2,
  sigmaby = 0.01,
  myfolds = 10,
  myseed1 = 123,
  myseed2 = 456
)
}
\arguments{
\item{mydf}{input data frame}

\item{myresponse}{the name of the dependent variable for the GPM model.  This will normally be the residuals of the OLS model: "residuals".}

\item{myterms}{a character vector containing the names of the independent variables.}

\item{sigmafrom}{the lower limit of the test range for sigma}

\item{sigmato}{the upper limit of the test range for sigma}

\item{sigmaby}{the increment of the test values for sigma}

\item{myfolds}{the number of folds for k-fold cross-validation}

\item{myseed1}{the first random number seed}

\item{myseed2}{the second random number seed}
}
\value{

}
\description{
Estimates a Gaussian Process Model.
}
\details{

}
\examples{



}
