% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ols.R
\name{model_ols}
\alias{model_ols}
\title{Ordinary Least Squares Model}
\usage{
model_ols(mydf = NULL, myresponse = NULL, myterms = NULL)
}
\arguments{
\item{mydf}{input data frame}

\item{myresponse}{the dependent variable}

\item{myterms}{a character vector containing the names of the independent variables.}
}
\value{

}
\description{
Estimates an Ordinary Least Squares model
}
\details{
The spreadsheet should have only one sheet with the following columns and one row for each variable to be simulated:
}
\examples{




}
