% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{mcr}
\alias{mcr}
\title{monte carlo simulation using mc2d}
\usage{
mcr(myfilename = NULL, myseed = 12345L, nsims = 1000)
}
\arguments{
\item{myfilename}{is the name of the input Excel file}

\item{myseed}{the random number seed to be used}
}
\value{

}
\description{
a wrapper for mc2d
}
\details{
The input spreadsheet should have only one sheet with the following columns and one row for each variable to be simulated:
\itemize{
\item Index (numeric)
\item Variable (character)
\item codename (character)
\item Base (numeric)
\item Minimum (numeric)
\item Mode (numeric)
\item Maximum (numeric)
\item Distribution (character) - Uniform, Triangular, or PERT.
\item Shape (numeric) - only needed if distribution is "PERT"
}
}
\examples{
myinputfile <- system.file("extdata", "risk_variable_distributions.xlsx", package = "humblr")
mytest1 <- mcr(myinputfile, nsims = 1000, myseed = 12345L)
mytest1

}
