% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ols_gpm.R
\name{model_ols_gpm}
\alias{model_ols_gpm}
\title{Model with OLS and GPM Components}
\usage{
model_ols_gpm(
  myinputspreadsheet = NULL,
  myinputdatasheet = NULL,
  mypredictiondf = NULL,
  mytag = NULL,
  mypredictionsheet = NULL,
  myresponse_ols = NULL,
  myresponse_gpm = "residuals",
  myterms = NULL,
  sigmafrom = 0.01,
  sigmato = 2,
  sigmaby = 0.01,
  myfolds = 10,
  myseed1 = 123,
  myseed2 = 456,
  mygraphsize = 1000
)
}
\arguments{
\item{myinputspreadsheet}{is the name of the spreadsheet with the input data.}

\item{myinputdatasheet}{the name of the sheet with the input data.}

\item{mypredictiondf}{(optional) a dataframe to be used as a basis for predictions}

\item{mytag}{a character string that will be used in the output file names.  #'}

\item{mypredictionsheet}{(optional) the name of a sheet containing test data to which the model should be applied.}

\item{myresponse_ols}{the name of the dependent variable for the OLS model}

\item{myresponse_gpm}{the name of the dependent variable for the GPM model.  This will normally be the residuals of the OLS model: "residuals".}

\item{myterms}{a character vector containing the names of the independent variables.}

\item{sigmafrom}{the lower limit of the test range for sigma}

\item{sigmato}{the upper limit of the test range for sigma}

\item{sigmaby}{the increment of the test values for sigma}

\item{myfolds}{the number of folds for k-fold cross-validation}

\item{myseed1}{the first random number seed}

\item{myseed2}{the second random number seed}

\item{mygraphsize}{the pixel dimension for the output graphs}
}
\value{

}
\description{
produces a model with Ordinary Least Squares (OLS) and Gaussian Process Model (GPM) components, based on input data provided in an Excel spreadsheet.
}
\details{
The spreadsheet requires at least one sheet, that must contain the dependent and independent variables.
}
\examples{
myinputfile <- system.file("extdata", "model_ols_gpm_test.xlsx", package = "humblr")
mytestmodel <- model_ols_gpm(myinputspreadsheet = myinputfile, myinputdatasheet = "data", mypredictionsheet = "prediction_tests", mytag = "model_ols_gpm_test", myresponse_ols = "depvar", myresponse_gpm = "residuals", myterms = c("var2",	"var3",	"var4",	"var5",	"var6",	"var7",	"var8",	"var9",	"var10"), sigmafrom = 1, sigmato= 10, sigmaby = 1, myfolds = 10, myseed1 = 123, myseed2 = 456, mygraphsize = 1000)


}
